Animations work and they take 16 mins to load but worth it
